refactor(entity): split cell_entity along struct/impl seams
Split the 1183-line crates/entity/src/cell_entity/mod.rs (>700-line hard
cap, CLAUDE.md "File organization") into cohesive sibling submodules:

- entity_struct.rs  — the CellEntity struct definition (irreducible single
                      type whose field set alone exceeds the soft cap)
- construction.rs   — CellEntity::new + the hand-rolled Debug projection
- witness_aoi.rs    — position get/set, witness-set mutators, is_in_aoi
- appearance.rs     — filter_holstered_weapon + holster/appearance methods
- weapon_action.rs  — clear_weapon_action_state pending-timer reset

mod.rs is now the type catalog (ActiveEffectInstance, NpcInteractionType,
LootItem, BandolierItem, AiState, MobMovementType) plus module wiring and
re-exports. All public items are re-exported (`pub use entity_struct::
CellEntity`, `pub use appearance::filter_holstered_weapon`, existing
`SystemOptions`), and the enums/aux structs remain defined in mod.rs, so
every `cimmeria_entity::cell_entity::*` path is unchanged.

The only tests.rs change is a path-only import fix (EntityId/SpaceId are
now pulled from cimmeria_common directly rather than glob-re-exported
through mod.rs). No logic or signature changes.

Part of #529, supersedes #535
